Implemented code that prevents an administrator from deleting an organization if the organization is currently associated with a user

+++ vms/organization/services.py
@@ -1,14 +1,24 @@
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 from organization.models import Organization
 
+#need to check that this organization is not currently associated with a user (otherwise the user gets cascade deleted)
 def delete_organization(organization_id):
 
+    result = True
+
     organization = get_organization_by_id(organization_id)
+    #check if there are currently any users associated with this organization
+    #this might be difficult to maintain as different types of users are added on
+    volunteers_in_organization = organization.volunteer_set.all()
+    administrators_in_organization = organization.administrator_set.all()
     
-    if organization:
+    #can only delete an organization if no users are currently associated with it
+    if organization and (not volunteers_in_organization) and (not administrators_in_organization):
         organization.delete()
     else:
-        return ObjectDoesNotExist
+        result = False
+
+    return result
 
 def get_organization_by_id(organization_id):
